# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: I.D.A Immobilier API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/register:
    post:
      operationId: registerClient
      tags: [auth]
      summary: Register a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /agencies:
    get:
      operationId: listAgencies
      tags: [agencies]
      summary: List all agencies
      responses:
        "200":
          description: Agencies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agency"
    post:
      operationId: createAgency
      tags: [agencies]
      summary: Create an agency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgencyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agency"
  /agencies/{id}:
    get:
      operationId: getAgency
      tags: [agencies]
      summary: Get agency by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agency
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agency"
    patch:
      operationId: updateAgency
      tags: [agencies]
      summary: Update an agency
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgencyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agency"
  /agencies/{id}/agents:
    get:
      operationId: listAgencyAgents
      tags: [agencies]
      summary: List agents for an agency
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /agencies/{id}/stats:
    get:
      operationId: getAgencyStats
      tags: [agencies]
      summary: Get agency analytics summary
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgencyStats"

  # USERS
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - in: query
          name: agentId
          schema:
            type: integer
        - in: query
          name: clientId
          schema:
            type: integer
        - in: query
          name: propertyId
          schema:
            type: integer
        - in: query
          name: from
          schema:
            type: string
            format: date
        - in: query
          name: to
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Cancel appointment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # FAVORITES
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RegisterInput:
      type: object
      required: [email, password, firstName, lastName, phone]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # USER SCHEMAS
    User:
      type: object
      required: [id, email, firstName, lastName, role, isActive, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [superadmin, admin, agency_manager, agent, client]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        agencyId:
          type: ["integer", "null"]
        agencyName:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Agency:
      type: object
      required: [id, name, email, phone, address, city, postalCode, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        postalCode:
          type: string
        website:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        agentCount:
          type: integer
        propertyCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    AgencyInput:
      type: object
      required: [name, email, phone, address, city, postalCode]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        postalCode:
          type: string
        website:
          type: string
        logoUrl:
          type: string
    AgencyUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        postalCode:
          type: string
        website:
          type: string
        logoUrl:
          type: string
        status:
          type: string
          enum: [active, inactive]
    AgencyStats:
      type: object
      required: [totalProperties, publishedProperties, totalLeads, totalAppointments, totalAgents]
      properties:
        totalProperties:
          type: integer
        publishedProperties:
          type: integer
        totalLeads:
          type: integer
        totalAppointments:
          type: integer
        totalAgents:
          type: integer
        avgPropertyPrice:
          type: ["number", "null"]

    # PROPERTY SCHEMAS
...
    Appointment:
      type: object
      required: [id, type, scheduledAt, status, agentId, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [visit, meeting, phone_call, other]
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        scheduledAt:
          type: string
          format: date-time
        durationMinutes:
          type: integer
        propertyId:
          type: ["integer", "null"]
        propertyTitle:
          type: ["string", "null"]
        propertyAddress:
          type: ["string", "null"]
        agentId:
          type: integer
        agentName:
          type: ["string", "null"]
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        clientEmail:
          type: ["string", "null"]
        clientPhone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AppointmentInput:
      type: object
      required: [type, scheduledAt, agentId]
      properties:
        type:
          type: string
          enum: [visit, meeting, phone_call, other]
        scheduledAt:
          type: string
          format: date-time
        durationMinutes:
          type: integer
        propertyId:
          type: integer
        agentId:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        clientEmail:
          type: string
        clientPhone:
          type: string
        notes:
          type: string
    AppointmentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        scheduledAt:
          type: string
          format: date-time
        durationMinutes:
          type: integer
        notes:
          type: string

    # FAVORITE SCHEMAS
...